requirements.txt
playwright
fastapi
uvicorn
pyyaml
pydantic